fire: Fix damage and ammunition use of return fire
quiet_bigdef() runs for each attacker.  It lets each eligible defender
fire at most once.  The first time a defender is eligible, it fires
and is saved in the list of defenders, along with its firing damage.
If it's eligible again for a later attacker, it's found in the list of
defenders, and the damage is reused.  The list of defenders searched
with search_flist().  Unfortunately, search_flist() compares only uid,
not type, and therefore can return a previously found defender of
another type.

If there are multiple attackers and multiple defenders with the same
uid, total damage can be off, damage can be spread to attackers out of
range, and defenders may not be charged shells.  Abuse is possible,
but complicated to set up, and probably not worth the trouble.

Broken in commit f89edc7, v4.3.12.  Fix by comparing the type as well.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>