power: Drop the RES_POP factor
If option RES_POP is enabled, the power factor is multiplied by a
"research factor" of 1.0 + maxpop / 10000.0, where maxpop is the
maximum population of a mine sector.

Back when this code was written (Chainsaw 3), all sectors had the same
population limit, so using a mine sector was as good as any.  Since
then, it has become configurable, and the stock game has both sector
types with lower (mountains, plains) and with higher (cities)
population limits.

Space for people is worth considering for power, but multiplying total
power by a fudge factor based on the most common sector type's maximum
population is silly.  Drop it.

Adjusting each sector's value for maximum population would make more
sense, with and without RES_POP.  Perhaps later.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>