bomb: Don't list bombable commodities on invalid selection
comm_bomb() first lists commodities that can be bombed and are
present, then prompts for the one to bomb.  If the player selects a
bomb-proof one, it rejects it, lists the bombable ones, and prompts
again.  This can only happen when option SUPER_BARS is enabled, and
the player selects bars.  Looks like this:

    Bomb what? (ship, plane, land unit, efficiency, commodities) c
    some civilians
    some military
    commodity to bomb? b
    You can't bomb bars of gold!
    Bombable: civilians, military, shells, guns, petrol, iron ore, dust (gold), food, oil, light products, heavy products, uncompensated workers, radioactive materials
    commodity to bomb?

The list is of marginal value.  It was more useful back when
comm_bomb() didn't list commodities before prompting (BSD Empire 1.1).
It's also illegible.  Drop it.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>