update: Don't allocate a bp map
The bp map is needed so that budget can track materials and work
available in sectors for ship, plane and land unit building without
updating the sector file.

For the update proper, going through the bp map is a pointless waste
of memory.  Use a null bp map there: writes become no-ops, and reads
read from the underlying sector.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>