update: Round the people's work randomly rather than down
Work has always been rounded down.  This can lead to a bit of work
micromanagement.  For instance, four military on a lonely island
accomplish nothing in 60 ETU updates, but five will make one point of
work per update.  They can build a 2% harbor in four updates, as long
as rollover_avail_max is at least 3.  Six to eight will be no faster.

Switch to random rounding, so that players don't have to get it
exactly right.  Four miliary now get to 2% in five updates on average,
five in four, six or seven in three, and so forth.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>