configure Make: More detailed version information
Version information is in output of commands version, xdump version,
and in program output for option -v.  Looks like this:

    Wolfpack Empire 4.3.33

It's defined in configure.ac, and incremented manually.  It identifies
only the base release (here: 4.3.33).  Fine when this is an unmodified
released version.  Version information doesn't tell you whether it is.

Add a suffix to the version that describes it further:

    V 	     	    Unmodified release V (same as before)
    V-dirty	    Modified release built from a tarball
    V.N-H	    Modified release built from a clean git tree
    		    N is the number of additional commits, and
		    H is the abbreviated commit hash
    V.N-H-dirty	    Same, but the working tree is dirty

A git tree is clean when the contents of its files are unchanged.
Changing only the their timestamps doesn't count.  It does count when
building from a tarball, because tracking contents isn't implemented.

Also use this suffixed version for tarballs.

configure's version is fixed at configure generation time, i.e. it is
usually out of date during development.  Ensuring a release tarball
gets a current one is manual for now.  Running autoconf -f should do
the trick.

Elsewhere, the version is determined at build time, so it is always
accurate.

Dirty tracking isn't implemented in the standalone client build.  If
you start with a clean tarball, the version will not change from V to
V-dirty when you build with modifications.

Steal build-aux/git-version-gen from autoconf 2.69 to help with
computing the version.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>