attack assault board: Fix message when attacking land unit dies
Loss of an attacking land is reported like this:

    @linf light infantry #1 dies assault 21,-3!

Bad grammar and newline missing between the "@" casualty character and
the message.  Messed up in Empire 2.  Affects only attackers, because
the code special-cases defense to avoid the bad grammar there (close,
but no cigar), and defenders don't get casualty characters printed.
Fix it to

    @
    linf light infantry #1 dies assaulting 21,-3!

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>